move tests to buffer pkg
